Moving migration to subclasses
