+++ src/modules/cmdpal/ext/Microsoft.CmdPal.Ext.TimeDate/Helpers/SettingsManager.cs
@@ -7,6 +7,8 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.IO;
 using System.Linq;
+using System.Text.Json;
+using System.Text.Json.Nodes;
 using Microsoft.CommandPalette.Extensions.Toolkit;
 
 namespace Microsoft.CmdPal.Ext.TimeDate.Helpers;
@@ -19,6 +21,11 @@ public class SettingsManager : JsonSettingsManager, ISettingsInterface
     private const string CUSTOMFORMATPLACEHOLDER = "MyFormat=dd-MMM-yyyy\rMySecondFormat=dddd (Da\\y nu\\mber: DOW)\rMyUtcFormat=UTC:hh:mm:ss";
 
     private static readonly string _namespace = "timeDate";
+
+    private static readonly JsonSerializerOptions _serializerOptions = new()
+    {
+        WriteIndented = true,
+    };
 
     private static string Namespaced(string propertyName) => $"{_namespace}.{propertyName}";
 
@@ -159,6 +166,58 @@ public class SettingsManager : JsonSettingsManager, ISettingsInterface
     {
         var directory = Utilities.BaseSettingsPath("Microsoft.CmdPal");
         return Path.Combine(directory, "settings.json");
+    }
+
+    /// <summary>
+    /// Migrates settings from a shared legacy file to this extension's own settings file.
+    /// Call after registering all settings with <see cref="Settings"/> and before <see cref="LoadSettings"/>.
+    /// Skips if <see cref="FilePath"/> already exists or <paramref name="legacyFilePath"/> is missing.
+    /// </summary>
+    private void MigrateFromLegacyFile(string legacyFilePath)
+    {
+        if (string.IsNullOrEmpty(FilePath) || string.IsNullOrEmpty(legacyFilePath))
+        {
+            return;
+        }
+
+        // Already migrated — per-extension file exists.
+        if (File.Exists(FilePath))
+        {
+            return;
+        }
+
+        if (!File.Exists(legacyFilePath))
+        {
+            return;
+        }
+
+        try
+        {
+            var legacyContent = File.ReadAllText(legacyFilePath);
+            if (JsonNode.Parse(legacyContent) is not JsonObject)
+            {
+                return;
+            }
+
+            // Extract only the keys this extension owns.
+            Settings.Update(legacyContent);
+            var settingsJson = Settings.ToJson();
+
+            if (JsonNode.Parse(settingsJson) is JsonObject extracted && extracted.Count > 0)
+            {
+                var directory = Path.GetDirectoryName(FilePath);
+                if (!string.IsNullOrEmpty(directory))
+                {
+                    Directory.CreateDirectory(directory);
+                }
+
+                File.WriteAllText(FilePath, extracted.ToJsonString(_serializerOptions));
+            }
+        }
+        catch (Exception ex)
+        {
+            ExtensionHost.LogMessage(new LogMessage() { Message = $"Settings migration failed from '{legacyFilePath}' to '{FilePath}': {ex}" });
+        }
     }
 
     public SettingsManager()
