+++ src/modules/cmdpal/extensionsdk/Microsoft.CommandPalette.Extensions.Toolkit/JsonSettingsManager.cs
@@ -18,58 +18,6 @@ public abstract class JsonSettingsManager
         WriteIndented = true,
     };
 
-    /// <summary>
-    /// Migrates settings from a shared legacy file to this extension's own settings file.
-    /// Call after registering all settings with <see cref="Settings"/> and before <see cref="LoadSettings"/>.
-    /// Skips if <see cref="FilePath"/> already exists or <paramref name="legacyFilePath"/> is missing.
-    /// </summary>
-    protected void MigrateFromLegacyFile(string legacyFilePath)
-    {
-        if (string.IsNullOrEmpty(FilePath) || string.IsNullOrEmpty(legacyFilePath))
-        {
-            return;
-        }
-
-        // Already migrated — per-extension file exists.
-        if (File.Exists(FilePath))
-        {
-            return;
-        }
-
-        if (!File.Exists(legacyFilePath))
-        {
-            return;
-        }
-
-        try
-        {
-            var legacyContent = File.ReadAllText(legacyFilePath);
-            if (JsonNode.Parse(legacyContent) is not JsonObject)
-            {
-                return;
-            }
-
-            // Extract only the keys this extension owns.
-            Settings.Update(legacyContent);
-            var settingsJson = Settings.ToJson();
-
-            if (JsonNode.Parse(settingsJson) is JsonObject extracted && extracted.Count > 0)
-            {
-                var directory = Path.GetDirectoryName(FilePath);
-                if (!string.IsNullOrEmpty(directory))
-                {
-                    Directory.CreateDirectory(directory);
-                }
-
-                File.WriteAllText(FilePath, extracted.ToJsonString(_serializerOptions));
-            }
-        }
-        catch (Exception ex)
-        {
-            ExtensionHost.LogMessage(new LogMessage() { Message = $"Settings migration failed from '{legacyFilePath}' to '{FilePath}': {ex}" });
-        }
-    }
-
     public virtual void LoadSettings()
     {
         if (string.IsNullOrEmpty(FilePath))
